requirements.txt
Django==3.2.5
djangorestframework==3.14.0
openai==0.26.4